--- domo/toy/src/main/java/com/domo/controller/UserCreateController.java
@@ -1,34 +1,0 @@
-package com.domo.controller;
-
-import com.domo.model.dto.UserCreateDto;
-import com.domo.model.dto.UserResponse;
-import com.domo.repository.UserEntity;
-import com.domo.service.UserService;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import io.swagger.v3.oas.annotations.tags.Tag;
-import lombok.RequiredArgsConstructor;
-
-@Tag(name = "유저(users)")
-@RestController
-@RequestMapping("/api/users")
-@RequiredArgsConstructor
-public class UserCreateController {
-
-    private final UserController userController;
-    private final UserService userService;
-
-    @PostMapping
-    public ResponseEntity<UserResponse> createUser(@RequestBody UserCreateDto userCreateDto) {
-        UserEntity userEntity = userService.create(userCreateDto);
-        return ResponseEntity
-            .status(HttpStatus.CREATED)
-            .body(userController.toResponse(userEntity));
-    }
-
-}
