--- domo/toy/src/main/java/com/domo/exception/CertificationCodeNotMatchedException.java
@@ -1,8 +1,0 @@
-package com.domo.exception;
-
-public class CertificationCodeNotMatchedException extends RuntimeException {
-
-    public CertificationCodeNotMatchedException() {
-        super("자격 증명에 실패하였습니다.");
-    }
-}
