--- domo/toy/src/main/java/com/domo/model/UserStatus.java
@@ -1,5 +1,0 @@
-package com.domo.model;
-
-public enum UserStatus {
-    PENDING, INACTIVE, ACTIVE
-}
