--- domo/toy/src/main/java/com/domo/model/dto/MyProfileResponse.java
@@ -1,17 +1,0 @@
-package com.domo.model.dto;
-
-import com.domo.model.UserStatus;
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class MyProfileResponse {
-
-    private Long id;
-    private String email;
-    private String nickname;
-    private String address;
-    private UserStatus status;
-    private Long lastLoginAt;
-}
