--- domo/toy/src/main/java/com/domo/model/dto/PostResponse.java
@@ -1,15 +1,0 @@
-package com.domo.model.dto;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class PostResponse {
-
-    private Long id;
-    private String content;
-    private Long createdAt;
-    private Long modifiedAt;
-    private UserResponse writer;
-}
