--- domo/toy/src/main/java/com/domo/model/dto/PostUpdateDto.java
@@ -1,18 +1,0 @@
-package com.domo.model.dto;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-public class PostUpdateDto {
-
-    private final String content;
-
-    @Builder
-    public PostUpdateDto(
-        @JsonProperty("content") String content) {
-        this.content = content;
-    }
-}
