--- domo/toy/src/main/java/com/domo/model/dto/UserResponse.java
@@ -1,16 +1,0 @@
-package com.domo.model.dto;
-
-import com.domo.model.UserStatus;
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class UserResponse {
-
-    private Long id;
-    private String email;
-    private String nickname;
-    private UserStatus status;
-    private Long lastLoginAt;
-}
