--- domo/toy/src/main/java/com/domo/model/dto/UserUpdateDto.java
@@ -1,21 +1,0 @@
-package com.domo.model.dto;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-public class UserUpdateDto {
-
-    private final String nickname;
-    private final String address;
-
-    @Builder
-    public UserUpdateDto(
-        @JsonProperty("nickname") String nickname,
-        @JsonProperty("address") String address) {
-        this.nickname = nickname;
-        this.address = address;
-    }
-}
