+++ domo/toy/src/main/java/com/domo/post/controller/response/PostResponse.java
@@ -1,0 +1,16 @@
+package com.domo.post.controller.response;
+
+import com.domo.user.controller.response.UserResponse;
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class PostResponse {
+
+    private Long id;
+    private String content;
+    private Long createdAt;
+    private Long modifiedAt;
+    private UserResponse writer;
+}
