+++ domo/toy/src/main/java/com/domo/post/domain/PostUpdate.java
@@ -1,0 +1,18 @@
+package com.domo.post.domain;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+public class PostUpdate {
+
+    private final String content;
+
+    @Builder
+    public PostUpdate(
+        @JsonProperty("content") String content) {
+        this.content = content;
+    }
+}
