+++ domo/toy/src/main/java/com/domo/post/infstructure/PostRepository.java
@@ -1,0 +1,7 @@
+package com.domo.post.infstructure;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface PostRepository extends JpaRepository<PostEntity, Long> {
+
+}
