+++ domo/toy/src/main/java/com/domo/post/service/PostService.java
@@ -1,0 +1,42 @@
+package com.domo.post.service;
+
+import java.time.Clock;
+
+import com.domo.common.domain.exception.ResourceNotFoundException;
+import com.domo.post.domain.PostCreate;
+import com.domo.post.domain.PostUpdate;
+import com.domo.post.infstructure.PostEntity;
+import com.domo.post.infstructure.PostRepository;
+import com.domo.user.infstructure.UserEntity;
+import com.domo.user.service.UserService;
+import org.springframework.stereotype.Service;
+
+import lombok.RequiredArgsConstructor;
+
+@Service
+@RequiredArgsConstructor
+public class PostService {
+
+    private final PostRepository postRepository;
+    private final UserService userService;
+
+    public PostEntity getPostById(long id) {
+        return postRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Posts", id));
+    }
+
+    public PostEntity create(PostCreate postCreate) {
+        UserEntity userEntity = userService.getById(postCreate.getWriterId());
+        PostEntity postEntity = new PostEntity();
+        postEntity.setWriter(userEntity);
+        postEntity.setContent(postCreate.getContent());
+        postEntity.setCreatedAt(Clock.systemUTC().millis());
+        return postRepository.save(postEntity);
+    }
+
+    public PostEntity update(long id, PostUpdate postUpdate) {
+        PostEntity postEntity = getPostById(id);
+        postEntity.setContent(postUpdate.getContent());
+        postEntity.setModifiedAt(Clock.systemUTC().millis());
+        return postRepository.save(postEntity);
+    }
+}
