--- domo/toy/src/main/java/com/domo/repository/PostRepository.java
@@ -1,7 +1,0 @@
-package com.domo.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface PostRepository extends JpaRepository<PostEntity, Long> {
-
-}
