--- domo/toy/src/main/java/com/domo/repository/UserRepository.java
@@ -1,14 +1,0 @@
-package com.domo.repository;
-
-import java.util.Optional;
-
-import com.domo.model.UserStatus;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-
-public interface UserRepository extends JpaRepository<UserEntity, Long> {
-
-    Optional<UserEntity> findByIdAndStatus(long id, UserStatus userStatus);
-
-    Optional<UserEntity> findByEmailAndStatus(String email, UserStatus userStatus);
-}
