--- domo/toy/src/main/java/com/domo/service/PostService.java
@@ -1,41 +1,0 @@
-package com.domo.service;
-
-import java.time.Clock;
-
-import com.domo.exception.ResourceNotFoundException;
-import com.domo.model.dto.PostCreateDto;
-import com.domo.model.dto.PostUpdateDto;
-import com.domo.repository.PostEntity;
-import com.domo.repository.PostRepository;
-import com.domo.repository.UserEntity;
-import org.springframework.stereotype.Service;
-
-import lombok.RequiredArgsConstructor;
-
-@Service
-@RequiredArgsConstructor
-public class PostService {
-
-    private final PostRepository postRepository;
-    private final UserService userService;
-
-    public PostEntity getPostById(long id) {
-        return postRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Posts", id));
-    }
-
-    public PostEntity create(PostCreateDto postCreateDto) {
-        UserEntity userEntity = userService.getById(postCreateDto.getWriterId());
-        PostEntity postEntity = new PostEntity();
-        postEntity.setWriter(userEntity);
-        postEntity.setContent(postCreateDto.getContent());
-        postEntity.setCreatedAt(Clock.systemUTC().millis());
-        return postRepository.save(postEntity);
-    }
-
-    public PostEntity update(long id, PostUpdateDto postUpdateDto) {
-        PostEntity postEntity = getPostById(id);
-        postEntity.setContent(postUpdateDto.getContent());
-        postEntity.setModifiedAt(Clock.systemUTC().millis());
-        return postRepository.save(postEntity);
-    }
-}
