+++ domo/toy/src/main/java/com/domo/user/controller/response/MyProfileResponse.java
@@ -1,0 +1,17 @@
+package com.domo.user.controller.response;
+
+import com.domo.user.domain.UserStatus;
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class MyProfileResponse {
+
+    private Long id;
+    private String email;
+    private String nickname;
+    private String address;
+    private UserStatus status;
+    private Long lastLoginAt;
+}
