+++ domo/toy/src/main/java/com/domo/user/domain/UserStatus.java
@@ -1,0 +1,5 @@
+package com.domo.user.domain;
+
+public enum UserStatus {
+    PENDING, INACTIVE, ACTIVE
+}
