+++ domo/toy/src/main/java/com/domo/user/infstructure/UserRepository.java
@@ -1,0 +1,14 @@
+package com.domo.user.infstructure;
+
+import java.util.Optional;
+
+import com.domo.user.domain.UserStatus;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+
+public interface UserRepository extends JpaRepository<UserEntity, Long> {
+
+    Optional<UserEntity> findByIdAndStatus(long id, UserStatus userStatus);
+
+    Optional<UserEntity> findByEmailAndStatus(String email, UserStatus userStatus);
+}
