--- domo/toy/src/test/java/com/domo/service/UserServiceTest.java
@@ -1,149 +1,0 @@
-package com.domo.service;
-
-import com.domo.exception.CertificationCodeNotMatchedException;
-import com.domo.exception.ResourceNotFoundException;
-import com.domo.model.UserStatus;
-import com.domo.model.dto.UserCreateDto;
-import com.domo.model.dto.UserUpdateDto;
-import com.domo.repository.UserEntity;
-import org.junit.jupiter.api.Test;
-import org.mockito.BDDMockito;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.mock.mockito.MockBean;
-import org.springframework.mail.SimpleMailMessage;
-import org.springframework.mail.javamail.JavaMailSender;
-import org.springframework.test.context.TestPropertySource;
-import org.springframework.test.context.jdbc.Sql;
-import org.springframework.test.context.jdbc.SqlGroup;
-
-import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
-import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
-import static org.junit.jupiter.api.Assertions.*;
-import static org.mockito.ArgumentMatchers.any;
-
-@SpringBootTest
-@TestPropertySource(properties = "spring.config.location=classpath:/application-test.yml")
-@SqlGroup({
-    @Sql("/sql/user-service-test-data.sql"),
-    @Sql(value = "/sql/delete-all-data.sql", executionPhase = Sql.ExecutionPhase.AFTER_TEST_METHOD)
-})
-
-class UserServiceTest {
-
-    @Autowired
-    private UserService userService;
-
-    @MockBean
-    private JavaMailSender mailSender;
-
-    @Test
-    void getByEmail은_ACTIVE_상태인_유저를_찾아올_수_있다() {
-        // given
-        String email = "me@dev-domo.com";
-
-        // when
-        UserEntity userEntity = userService.getByEmail(email);
-
-        // then
-        assertThat(userEntity.getNickname()).isEqualTo("domo");
-    }
-
-    @Test
-    void getByEmail은_PENDING_상태인_유저는_찾아올_수_없다() {
-        // given
-        String email = "me@dev-domo1.com";
-
-        // when, then
-        assertThatThrownBy(() -> userService.getByEmail(email))
-            .isInstanceOf(ResourceNotFoundException.class);
-    }
-
-    @Test
-    void getById는_ACTIVE_상태인_유저를_찾아올_수_있다() {
-        // given
-        // when
-        UserEntity userEntity = userService.getById(1);
-
-        // then
-        assertThat(userEntity.getNickname()).isEqualTo("domo");
-    }
-
-    @Test
-    void getById는_PENDING_상태인_유저는_찾아올_수_없다() {
-        // given
-        // when, then
-        assertThatThrownBy(() -> userService.getById(2))
-                .isInstanceOf(ResourceNotFoundException.class);
-    }
-
-    @Test
-    void userCreateDto를_이용하여_유저를_생성할_수_있다() {
-        // given
-        UserCreateDto userCreateDto = UserCreateDto.builder()
-                .email("me@dev-domo2.com")
-                .address("Seoul")
-                .nickname("domo2")
-                .build();
-        BDDMockito.doNothing().when(mailSender).send(any(SimpleMailMessage.class));
-
-        // when
-        UserEntity result = userService.create(userCreateDto);
-
-        // then
-        assertThat(result.getId()).isNotNull();
-        assertThat(result.getStatus()).isEqualTo(UserStatus.PENDING);
-//        assertThat(result.getCertificationCode()).isEqualTo("T^T");
-    }
-
-    @Test
-    void userUpdateDto를_이용하여_유저를_수정할_수_있다() {
-        // given
-        UserUpdateDto userUpdateDto = UserUpdateDto.builder()
-                .address("Inchon")
-                .nickname("domo짱")
-                .build();
-        BDDMockito.doNothing().when(mailSender).send(any(SimpleMailMessage.class));
-
-        // when
-        UserEntity result = userService.update(1, userUpdateDto);
-
-        // then
-        UserEntity userEntity = userService.getById(1);
-        assertThat(result.getId()).isNotNull();
-        assertThat(result.getAddress()).isEqualTo("Inchon");
-        assertThat(result.getNickname()).isEqualTo("domo짱");
-    }
-
-    @Test
-    void user를_로그인_시키면_마지막_로그인_시간이_변경된다() {
-        // given
-
-        // when
-        userService.login(1);
-
-        // then
-        UserEntity userEntity = userService.getById(1);
-        assertThat(userEntity.getLastLoginAt()).isGreaterThan(0L);
-//        assertThat(userEntity.getLastLoginAt()).isGreaterThan(T^T);
-    }
-
-    @Test
-    void PENDING상태인_사용자는_인증코드로_ACTIVE_시킬_수_있다() {
-        // given
-        // when
-        userService.verifyEmail(2, "aaaaaaaa-aaaa-aaaa-aaaa-aaaaaaaaaaab");
-
-        // then
-        UserEntity userEntity = userService.getById(2);
-        assertThat(userEntity.getStatus()).isEqualTo(UserStatus.ACTIVE);
-    }
-
-    @Test
-    void PENDING상태인_사용자는_잘못된_인증코드를_받으면_에러가_발생한다() {
-        // given
-        // when, hen
-        assertThatThrownBy(() -> userService.verifyEmail(2, "aaaaaaaa-aaaa-aaaa-aaaa-aaaaaaaaaaac"))
-                .isInstanceOf(CertificationCodeNotMatchedException.class);
-    }
-}
