--- domo/toy/src/test/resources/sql/post-create-controller-test-data.sql
@@ -1,2 +1,0 @@
-insert into `users` (`id`, `email`, `nickname`, `address`, `certification_code`, `status`, `last_login_at`)
-values (1, 'me@dev-domo.com', 'domo', 'Seoul', 'aaaaaaaa-aaaa-aaaa-aaaa-aaaaaaaaaaaa', 'ACTIVE', 0);
